requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers==3.0.1
pydantic==2.11.3
python-dotenv==1.0.1
